Software/Controller/requirements.txt
Adafruit_GPIO
simple-pid
RPi.GPIO
blynklib
adafruit-circuitpython-charlcd